Move the clock prompt to the progress pipeline file
